Moved page manipulation functions
The functions to create the header, nav and footer, which are elements present in all pages, were moved from the home module to the index module.
Also added event listeners to the nav buttons to change the page according to the button pressed.
Added a clearDOM function to clear the DOM whenever the user switches tabs.